:sparkles: 포스트 삭제기능 구현
세션을 확인하고 작성자와 일치하면 포스트 삭제버튼이 렌더링되고
삭제요청을 보냅니다

+++ pages/api/posts/[id].ts
@@ -1,6 +1,8 @@
 import prisma from "@/libs/server/prisma";
 import withHandler from "@/libs/server/withHandler";
 import { NextApiRequest, NextApiResponse } from "next";
+import { getServerSession } from "next-auth";
+import { authOptions } from "../auth/[...nextauth]";
 
 async function handler(req: NextApiRequest, res: NextApiResponse) {
   const { id } = req.query;
@@ -8,21 +10,29 @@ async function handler(req: NextApiRequest, res: NextApiResponse) {
   if (!id) {
     return res.status(400).json({ error: "id를 찾을 수 없습니다" });
   } else {
-    if (req.method === "DELETE") {
-      await prisma.post.delete({ where: { id: id.toString() } });
-      res.status(200).json({ message: "포스트가 삭제되었습니다" });
-    }
-    const post = await prisma.post.findUnique({
-      where: { id: id.toString() },
-      include: {
-        user: {
-          select: { id: true, name: true, image: true },
+    if (req.method === "GET") {
+      const post = await prisma.post.findUnique({
+        where: { id: id.toString() },
+        include: {
+          user: {
+            select: { id: true, name: true, image: true },
+          },
         },
-      },
-    });
-    console.log(post);
-    res.json(post);
+      });
+      res.json(post);
+    }
+
+    const session = await getServerSession(req, res, authOptions);
+    if (session) {
+      if (req.method === "DELETE") {
+        await prisma.post.delete({ where: { id: id.toString() } });
+        res.status(200).json({ message: "포스트가 삭제되었습니다" });
+      }
+      // if(req.method === "POST"){
+      //   const alreadyExists = Boolean(await prisma.user.findUnique({where:email}))
+      // }
+    }
   }
 }
 
-export default withHandler(["GET", "DELETE"], handler);
+export default withHandler(["GET", "DELETE", "POST"], handler);
